Strip unnecessary changes
