+++ packages/playwright/src/reporters/list.ts
@@ -258,7 +258,7 @@ class ListReporter extends TerminalReporter {
   override async onEnd(result: FullResult) {
     await super.onEnd(result);
     process.stdout.write('\n');
-    await this.epilogue(true);
+    this.epilogue(true);
   }
 }
